test: add test for /vault
