+++ sveltekit/src/lib/helpers/accessKey.server.ts
@@ -1,5 +1,5 @@
 import prisma from '$lib/helpers/prisma';
-import { AccessKeySettings } from '$lib/helpers/constants';
+import { AccessKeySettings, ACCESS_KEY_COOKIE_NAME } from '$lib/helpers/constants';
 import type { AccessKeyStatusStore } from '$lib/stores/accessKeyStatusStore';
 
 export const getAccessKeySettings = async (): Promise<AccessKeyStatusStore> => {
@@ -15,4 +15,19 @@ export const getAccessKeySettings = async (): Promise<AccessKeyStatusStore> => {
 	return {
 		accessKey: accessKey?.value
 	};
+};
+
+export const getVaultAccessKey = async () => {
+	return await prisma.setting.findUnique({
+		where: { name: AccessKeySettings.ACCESS_KEY },
+		select: { value: true }
+	});
+};
+
+export const isRequestAuthorized = async (request: Request) => {
+	const cookie = request.headers.get('cookie');
+	const requestAccessKey = cookie?.split(ACCESS_KEY_COOKIE_NAME)[1];
+	const vaultAccessKey = await getVaultAccessKey();
+
+	return vaultAccessKey === null || vaultAccessKey?.value === requestAccessKey;
 };
